services/inventory_service/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.16
django-filter>=24.3,<25.0
drf-spectacular>=0.27,<0.28
psycopg[binary]>=3.2,<3.3
gunicorn>=22.0,<23.0
python-dotenv>=1.0,<2.0
redis>=5.0,<6.0
celery>=5.4,<5.5
httpx>=0.28,<0.29
neo4j>=5.26,<6.0
shortuuid>=1.0,<2.0
